test/python/requirements.txt
pip==23.0
wheel~=0.38.4
setuptools==67.3.2
requests~=2.28.2
websocket-client~=1.5.1
packaging~=23.0
docker~=6.0.1
certifi~=2022.12.7
idna~=3.4
urllib3~=1.26.14
h11~=0.14.0
attrs~=22.2.0
anyio~=3.6.2
sniffio~=1.3.0
httpcore~=0.16.3
rfc3986==2.0.0
httpx~=0.23.3